fix: refactoring code
