+++ src/App.css
@@ -25,157 +25,4 @@ a {
   align-items: center;
   background-color: #FF3CAC;
   background-image: linear-gradient(225deg, #FF3CAC 0%, #784BA0 50%, #2B86C5 100%);
-
-}
-
-
-.game-field {
-  width: 500px;
-  height: 500px;
-  border: 1px solid white;
-  position: relative;
-  /*background: #34e89e; */ /* fallback for old browsers */
- /* background: -webkit-linear-gradient(to right, #0f3443, #34e89e);*/  /* Chrome 10-25, Safari 5.1-6 */
-  /*background: linear-gradient(to right, #0f3443, #34e89e);*/ /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
-  background-color: rgba(0, 0, 0, 0.2);
-}
-
-.game-field.default {
-  border: 1px solid black;
-  background-color: rgba(255, 255, 255, 0.2);
-}
-.snake-body {
-  position: absolute;
-  width: 2%;
-  height: 2%;
-  background-color: black;
-  border: 1px solid white;
-  z-index: 2;
-}
-
-.snake-body.default {
-  background-color: white;
-  border: 1px solid white;
-}
-
-.food {
-  width: 2%;
-  height: 2%;
-  background-color: red;
-  position: absolute;
-  z-index: 1;
-}
-.log{
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  padding: 25% 0 0 0;
-  z-index: 5;
-  text-align: center;
-  font-size: 25px;
-  background-color: rgba(255, 255, 255, 0.2);
-}
-.log.none {
-  display: none;
-}
-.game-cotainer {
-  width: 500px;
-}
-
-.panel {
-  height: 50px;
-  display: flex;
-  text-align: center;
-  justify-content: space-around;
-}
-.panel button {
-  height: 25px;
-  width: 75px;
-  background-color: rgba(0, 0, 0, 0.3);
-  outline: none;
-}
-
-.lds-roller {
-  display: inline-block;
-  position: relative;
-  width: 80px;
-  height: 80px;
-}
-.lds-roller div {
-  animation: lds-roller 1.2s cubic-bezier(0.5, 0, 0.5, 1) infinite;
-  transform-origin: 40px 40px;
-}
-.lds-roller div:after {
-  content: " ";
-  display: block;
-  position: absolute;
-  width: 7px;
-  height: 7px;
-  border-radius: 50%;
-  background: #fff;
-  margin: -4px 0 0 -4px;
-}
-.lds-roller div:nth-child(1) {
-  animation-delay: -0.036s;
-}
-.lds-roller div:nth-child(1):after {
-  top: 63px;
-  left: 63px;
-}
-.lds-roller div:nth-child(2) {
-  animation-delay: -0.072s;
-}
-.lds-roller div:nth-child(2):after {
-  top: 68px;
-  left: 56px;
-}
-.lds-roller div:nth-child(3) {
-  animation-delay: -0.108s;
-}
-.lds-roller div:nth-child(3):after {
-  top: 71px;
-  left: 48px;
-}
-.lds-roller div:nth-child(4) {
-  animation-delay: -0.144s;
-}
-.lds-roller div:nth-child(4):after {
-  top: 72px;
-  left: 40px;
-}
-.lds-roller div:nth-child(5) {
-  animation-delay: -0.18s;
-}
-.lds-roller div:nth-child(5):after {
-  top: 71px;
-  left: 32px;
-}
-.lds-roller div:nth-child(6) {
-  animation-delay: -0.216s;
-}
-.lds-roller div:nth-child(6):after {
-  top: 68px;
-  left: 24px;
-}
-.lds-roller div:nth-child(7) {
-  animation-delay: -0.252s;
-}
-.lds-roller div:nth-child(7):after {
-  top: 63px;
-  left: 17px;
-}
-.lds-roller div:nth-child(8) {
-  animation-delay: -0.288s;
-}
-.lds-roller div:nth-child(8):after {
-  top: 56px;
-  left: 12px;
-}
-@keyframes lds-roller {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
 }
