+++ src/components/Header/header.css
@@ -2,6 +2,8 @@
     width: 100%;
     height: 50px;
     display: flex;
+    justify-content: space-between;
+    padding: 0px 20px;
 }
 .back-arrow {
     width: 50px;
