+++ src/components/Navigation/navigation.css
@@ -2,14 +2,12 @@
     font-size: 25px;
     text-align: center;
 }
+
 .navigation li {
     padding: 5px 0px;
 }
-/*.navigation li a {
-    text-shadow: 1px 1px 2px white, 0 0 1em white;
-}*/
 
-/*.navigation li */a:hover {
+a:hover {
     text-shadow: 3px 3px rgb(0, 0, 0, 0.3);
     transition: 0.2;
     color: black;
