+++ src/components/game/StartGame/Log/log.css
@@ -1,0 +1,14 @@
+.log{
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    padding: 25% 0 0 0;
+    z-index: 5;
+    text-align: center;
+    font-size: 25px;
+    background-color: rgba(255, 255, 255, 0.2);
+}
+
+.log.none {
+    display: none;
+}
