+++ src/components/game/StartGame/Snake/Snake.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import './snake.css';
 
 function Snake(props) {
     return (
