+++ src/components/game/StartGame/startGame.css
@@ -1,0 +1,32 @@
+.game-field {
+    width: 500px;
+    height: 500px;
+    border: 1px solid white;
+    position: relative;
+    /*background: #34e89e; */ /* fallback for old browsers */
+   /* background: -webkit-linear-gradient(to right, #0f3443, #34e89e);*/  /* Chrome 10-25, Safari 5.1-6 */
+    /*background: linear-gradient(to right, #0f3443, #34e89e);*/ /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
+    background-color: rgba(0, 0, 0, 0.2);
+}
+  
+.game-field.default {
+    border: 1px solid black;
+    background-color: rgba(255, 255, 255, 0.2);
+}
+  
+.game-cotainer {
+    width: 500px;
+}
+  
+.panel {
+    height: 50px;
+    display: flex;
+    text-align: center;
+    justify-content: space-around;
+}
+.panel button {
+    height: 25px;
+    width: 75px;
+    background-color: rgba(0, 0, 0, 0.3);
+    outline: none;
+}
