--- src/index.css
@@ -1,14 +1,0 @@
-.game-field {
-  width: 500px;
-  height: 500px;
-  border-radius: 1px solid black;
-  position: relative;
-  /*margin: px auto;*/
-}
-.snake {
-  position: absolute;
-  width: 2%;
-  height: 2%;
-  background-color: black;
-  border: 1px solid black;
-}
